requirements.txt
configparser==4.0.2
prettytable==0.7.2
python-nmap==0.6.1
